Move interactor package into internal root
